requirements.txt
numpy 
pandas 
scikit-learn
matplotlib
seaborn
pymysql
mysql-connector-python
python-dotenv